+++ src/domain/ticker/specs/ticker.controller.spec.ts
@@ -1,19 +1,19 @@
 import { HttpModule } from '@nestjs/axios'
 import { Test, TestingModule } from '@nestjs/testing'
-import { TickerRequestDto, TickerResponseDto } from './dto'
-import { TickerController } from './ticker.controller'
-import { TickerService } from './ticker.service'
+import { TickerRequestDto, TickerResponseDto } from './../dto'
+import { TickerController } from './../ticker.controller'
+import { TickerService } from './../ticker.service'
 
-import { BinanceRepository } from './repositories'
-import { BitfinexRepository } from './repositories'
-import { BitgetRepository } from './repositories'
-import { BitstampRepository } from './repositories'
-import { CoinbaseRepository } from './repositories'
-import { GateIORepository } from './repositories'
-import { KrakenRepository } from './repositories'
-import { KuCoinRepository } from './repositories'
-import { MercadoBitcoinRepository } from './repositories'
-import { OKXRepository } from './repositories'
+import { BinanceRepository } from './../repositories'
+import { BitfinexRepository } from './../repositories'
+import { BitgetRepository } from './../repositories'
+import { BitstampRepository } from './../repositories'
+import { CoinbaseRepository } from './../repositories'
+import { GateIORepository } from './../repositories'
+import { KrakenRepository } from './../repositories'
+import { KuCoinRepository } from './../repositories'
+import { MercadoBitcoinRepository } from './../repositories'
+import { OKXRepository } from './../repositories'
 
 const RepositoryMock = {
   getTicker({ symbol }): TickerResponseDto {
