+++ src/domain/ticker/specs/ticker.service.spec.ts
@@ -1,17 +1,17 @@
 import { HttpModule } from '@nestjs/axios'
 import { Test, TestingModule } from '@nestjs/testing'
-import { TickerService } from './ticker.service'
+import { TickerService } from './../ticker.service'
 
-import { BinanceRepository } from './repositories'
-import { BitfinexRepository } from './repositories'
-import { BitgetRepository } from './repositories'
-import { BitstampRepository } from './repositories'
-import { CoinbaseRepository } from './repositories'
-import { GateIORepository } from './repositories'
-import { KrakenRepository } from './repositories'
-import { KuCoinRepository } from './repositories'
-import { MercadoBitcoinRepository } from './repositories'
-import { OKXRepository } from './repositories'
+import { BinanceRepository } from './../repositories'
+import { BitfinexRepository } from './../repositories'
+import { BitgetRepository } from './../repositories'
+import { BitstampRepository } from './../repositories'
+import { CoinbaseRepository } from './../repositories'
+import { GateIORepository } from './../repositories'
+import { KrakenRepository } from './../repositories'
+import { KuCoinRepository } from './../repositories'
+import { MercadoBitcoinRepository } from './../repositories'
+import { OKXRepository } from './../repositories'
 
 describe('TickerService', () => {
   let service: TickerService
